Adding dependency for testing
To be able to do a test of the inflict damage function, we need to build a test environment that simulates the world to check if the function is actually registering the changes and updating the mercenary.
This environment is achieved by using dependency which must be installed on the machine using asdf.
Sources:
https://book.dojoengine.org/framework/testing